setup prisma and cockroachdb
