+++ packages/server/package.json
@@ -9,6 +9,7 @@
     "build": "tsc && mv dist/src/* dist/"
   },
   "dependencies": {
+    "@prisma/client": "3.12.0",
     "@trpc/server": "^9.20.3",
     "dotenv": "^16.0.0",
     "google-spreadsheet": "^3.2.0",
@@ -17,8 +18,10 @@
   },
   "devDependencies": {
     "@types/google-spreadsheet": "^3.2.0",
+    "@types/node": "^17.0.23",
     "@types/ws": "^8.5.3",
     "nodemon": "^2.0.15",
+    "prisma": "3.12.0",
     "ts-node": "^10.7.0",
     "typescript": "^4.6.2"
   }
